Catch C-g
When the prompt for pause appears, I'm often in the middle of some
other mini-buffer operation. When hitting C-g, it happens that I
inadvertently abort the pomodoro-prompt, leaving pomodoro in an
undefined state (not sure in which state actually). This change
catches C-g and lets the user resume pomodoro when ready.